Deleting Todos by User ID

+++ 04-Project3-TodoApp/routers/todos.py
@@ -72,10 +72,14 @@ async def update_todo(user: user_dependency, db: db_dependency, todo_request: To
 
 # Deleting Data
 @router.delete(path="/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication Failed")
+
+    todo_model = (db.query(Todos).filter(Todos.id == todo_id)
+                  .filter(Todos.owner_id == user.get("id")).first())
     if todo_model is None:
         return HTTPException(status_code=404, detail="Todo Not Found")
 
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("id")).delete()
     db.commit()
